Make ByteString.toString() to return a summary instead of the whole array as an Avro string.
Now, ByteString.toString() returns a summary that includes at most the
first and last four bytes (as hex values) and the total length.  Since
we don't know what character encoding is used, or even whether the
bytes represent an encoded string, we should not assume any encoding
when toString() is called.  If the user wants to interpret the
ByteString as an encoded character string, he should call one of the
other methods that specifies an encoding.  Also, including the whole
ByteString in the toString() is not optimal since the toString() may
end up in log files.

RB=121895
R=si-dev,jbetz,cpettitt,slim,dyoung
A=jbetz,cpettitt,slim,dyoung