Adding the ability to specify defaultRouting = Master in the config for Rest.li services.
In a multi-colo environment this allows you to set service-by-service whether it
should route to the default colo (local) or to the master colo (maybe local). We're
adding this for single-master services so that client code does not need to explicitly
route Rest.li requests to the master colo in each request.

Example config for a service in config/app/d2-config-cmdline/application.src:

<entry key="premiumServices">
  <map>
    <entry key="path" value="/premiumServices" />
    <entry key="loadBalancerStrategyProperties" value="${paymentsStrategyProperties}"/>
    <entry key="defaultRouting" value="Master"/> <!-- new -->
  </map>
</entry>

RB=179212