Complex Key singleton array list.
Previously, complex keys with array types could fail to be read correctly on the server side if it was used in a get
request and the array has only a single element in it.  This fixes that and adds tests to make sure that Models with
array types are translated correctly to the server even when they only have a single element.

RB=195243
BUGS=SI-873
R=osumampo,jbetz,si-dev,gogenc
A=gogenc