Providing a way to get the response payload and status without catching exceptions in case of a Rest.Li error.
Added hasError to Response and hasDecodedResponse to RestLiResponseException. Changed the logic to throw an exception if the error response cannot be decoded inside a RestLiResponseException.

Moved the error handling control from the ResponseFuture onto RestClient in order to be better positioned for a future change to do a similar control mechanism for callback based requests as well as for parseq client.

Adding changelog and gradle.properties changes.