Check cache contents for race condition
There is a subtle race case in the code where, due to sluggish zookeeper, the
getStrategy() function returns empty list because the strategies are being
added. This causes problem in getStrategiesForService() because it caches
the original results, which is empty.

This change adds a check for the cache contents. Since each service has its
strategy, we force to re-getStrategy() even if the cache exists but there
is no entries.

Bug SI-1915

RB=462831
BUGS=SI-1915
R=dhoa,zzhu,axu
A=zzhu