Notify listeners when SimpleLoadBalancerState shutdown
If SimpleLoadBalancerState is shutdown, all the strategies and tracker
clients for each service are actually removed. This requires to notify
the state listeners if they need to take actions for the removals.

We believe this is the reason for duplicate sensor registration
because zookeeper session expiration need shutdown and restart the state.

RB=636452
BUG=SI-2239
G=si-core-reviewers
R=dhoa,zzhu,axu
A=ssheng