Improve memory and performance efficiency for LoadBalancerState
Currently when DegraderLoadBalancerStrategy updates state, it always builds up
a new consistent hashing ring with thousands of new points, therefore not most
efficient. This change tries to address this problem in two ways:

1. Reuse all the points when possible. A separate ring factory keeps all the
   generated points around, and only create new ones when needed. This can
   reduce the garbage generating and avoid the computational intensive hashing
   calculation. The original idea came from Diego Buthay in RB 570649.

2. Replace the TreeSet + arrays of objects with ArrayList to achieve the same
   performance with less memory consumption therefore GC friendly.

fix the dirty bits in the hasing in ConsistentHashRing, as found by Andrew Carter.

The orignal implementation was done by David (RB 646413). Updated with Steve's input.

Orignal comments:

I did a simple diff of the highest area percentage and the lowest area percentage to guage whether this is an improvement. The assert in the testcase verifies this. With the fix, the difference was 1.53, and without the fix it was 1.69. Obviously, since this testcase was only dealing with 10 values, these results may not be reflective of prod usage, but it should be good enough to catch regressions.
refactored the coverageMap code to be reusable between printRingArea() and my new test.

on line 289, I fixed a findBug warning about autoBoxing and immediately unautoboxing.

the two testcases in SimpleLoadBalancer were relying on a specific order in the lists, which changed with this fix. I made the tests not reliant on order.

impact of this change (please verify my thinking): as clients pick up this new version of pegasus, those clients should be sending traffic MORE evenly across clusters, so I don't believe it will cause any capacity issues, as all hosts currently have to be able to support the high end of the current skew. it will temporarily cause requests that use d2 stickiness for requests to reroute, but that's no different than when nodes are bounced. If these are the only impacts, then I'll alert pegasus-users and SRE's before rolling out, but I don't think we need anything from our users to roll this out. Let me know if you think otherwise.