Isolation of gradle's classpath from pegasus plugin's classpath.
This is a follow up on RB# 742894 and RB# 677995.

- Change from RB# 742894: for GenerateAvroSchemaTask.groovy, choose not to input avroTranslateOptionalDefault if it is null (as it leads to compilation error for a couple of MPs during container PCL).

- Possible impact: MPs that use gradle-plugins of pegasus, but most of them should be fine as they use it in the toolchain section. Only 3 MPs use the gradle-plugins in the product section: resource-identity, ligradle-pegasus and client-typed-urn-generator. From local mint build, only resource-identity requires change due to the pegasus API change. RB will be submitted to resource-identity to pick up this RB.

- Details of the motivation (as stated in RB# 742894):

The driver behind this is that there is no isolation
between gradle's classpath and the classpath of the pegasus
plugin. Since the plugin relies on things not being on the
classpath, we need better isolation. An added benefit with
this change is that the plugin dependencies won't be required
to be present at runtime. This means that when you load
the pegasus plugin it won't have to resolve everything
that it uses and we can delay the resolution until later.

Another benifit is that by the version being part of
the project configurations it will then be part of
our forcing strategy.

RB=815196
G=si-dev
R=ethall,kjin,dhoa,xma,dmajumde
A=kjin