Fix the racing condition for healthcheckMap map
HealthCheckMap is shared among paratitions so there is a potential race conditition that one thread updating
the map while another one is reading it. For the same reason HealthChecClientBuilder cannot be shared unless
it is protected by lock. A new builder is created whenever a new healthCheckClient is needed. This is a
tradoff between complicated implementation vs some young gen garbage.

RB=855000
G=si-core-reviewers
R=ssheng,dhoa,axu
A=ssheng