Add attribute pagingSupported in method schema and indicate in restspec if get_all and finder methods support paging.
The feature was requested by partner engineering team who wanted to
indicate to external developers if a method supported paging.

The solution was to look at method parameters of finder and get_all
methods and if it contained PagingContextParam, then add an attribute to
the idl to indicate the method supported paging.
This change has the side effect of generating modified idls for any
existing resource that had a get_all or finder method with paging. Hence
it was decided to make this change a major version so current users do
not see un-intended changes to rest specs.

Backward Compatibility:
The pegasus backwards compatibility checker is updated to include the
pagingSupported field during the checks.
  - Adding paging support to existing method is backward compatible.
  - Removing pagain support to existing method is backward incompatible.
    This is because a client might have application logic relying on the
servers pagination support and removing it can cause undesired
side-effects.

RB=885516
G=si-dev
R=aponniah,xma,mnchen
A=aponniah,kvidhani