Force to convert QuarantineLatency to String
Without the forcing, it ends up with a number when used to serialize the properties for ZK writes. This is inconsistent to d2-config-cmdline original behavior and causes problem when d2-config-service tries to de-serialize it back because the number is always converted to Integer instead of Long type.

RB=896742
G=si-core-reviewers
R=ssheng,dhoa
A=dhoa