[pegasus] Modify JacksonDataCodec to use FastByteArrayOutputStream for serialization and de-serialization.
FastByteArrayOutputStream is an alternative of java.io.ByteArrayOutputStream, and it uses a linked list of byte arrays for internal storage.
The advantage over java.io.ByteArrayOutputStream is that FastByteArrayOutputStream does not require to copy buffer or discard original array on resize, which uses less memory.
FastByteArrayOutputStream is derived from org.springframework.util.FastByteArrayOutputStream.

RB=937637
G=si-dev
R=kbalasub,sihde,mnchen,dhoa,xma,kvidhani,ssheng
A=kbalasub,mnchen