Change the quarantine health checking latency to use degrader low latency
Degrader low latency reflects the threshold for the service to be considered
as health. Use this value for quarantine health checking as well. The orignal
property 'http.loadBalancer.quarantine.latency' is removed as well.

Also fix the 'null' reason problem for play applications when ServiceUnavaliableException
thrown.

RB=958964
G=si-core-reviewers
R=ssheng,bfeng,dhoa
A=ssheng,dhoa