Improved flexibility and configurability
Improved flexibility and configurability of GenerateDataTemplateTask task
needed for Play on Gradle / voyager-api, while keeping backwards compatibility.

### Motivation

While working on Play on Gradle support for voyager-api
we found it necessary to convert GenerateDataTemplateTask to JavaExec task.
Doing so offers more configurability and flexibility
because we can leverage JavaExec task API to tweak the behavior of the task.
For example, clients can tweak jvm args of the task
and use 'li-pathing-jar' plugin to tackle command line arg limit for classpath.

### Implementation

1. In constructor we additionally create a 'config task' that is responsible for configuring the instance of GenerateDataTemplateTask.
See the code comment that describes the details.
2. Kept backwards binary compatibility of all the public properties of the task.
Type hierarchy is changed due to the change of the base class
but it should not be problematic because clients do not extend GenerateDataTemplateTask class but only refer to the type / task or create instances.
3. Started using Groovy's @CompileStatic to catch problems earlier
during compilation time. It also makes code execution faster.
4. Tested by running the build and inspecting the behavior of the config task and the target task.

RB=976188
BUG=TOOLS-143814
G=si-core-reviewers,gradle-play-reviewers
R=gylin,kbalasub
A=kbalasub