4th step of plan for sharing ChannelPoolManager with several Clients (https://docs.google.com/document/d/1aWqWRFvW5hjrcdU58oi0CqFe_lr4ial_lRqGFN-xRfc/edit)
Previously each Client would have had it's own Sensor on the ChannelPoolManager to aggregate stats on the processes.
To achieve sharing of Clients between services, the ChannelPoolManager will be shared and therefore also the underlying sensor.

Therefore the Sensor won't be identified from the service name anymore, but with a combination of pool stats prefix (new property HttpClientFactory.HTTP_POOL_STATS_NAME_PREFIX) and a hash of all the transport client properties that identify a ChannelPoolManager (and therefore also a ChannelPoolManagerBuilder)

This change comes with a relative change in the container's tests.

The next step is starting implementing the sharing logic.

RB=991779
G=si-core-reviewers
R=ssheng,dhoa
A=ssheng,dhoa