Adding 'fast recovery' mode to help host recover in low QPS situation
When average QPS is low for individual host, it is generally harder for the host to recover from
bad state because it has less chance to get the requests. RecoveryMap comes to help by increasing
the host points even though its callCount is 0. However the host is removed whenever it gets traffic once.
Therefore it can stay in bad state for long time if its point is still low.

This change tries to keep the host in recoveryMap for longer time even if it get traffic, until its point reaches the threshold.

RB=1048086
G=si-core-reviewers
R=ssheng,bfeng,dhoa
A=ssheng,dhoa