RELEASING 14.0.9
In the http2 specification referenced below, is suggested that the upgrade request clients use OPTIONS request for path “*” or a HEAD request for “/”. We are currently using the path of the actual user request that will be made after the upgrade (_path variable).

Therefore in the current implementation, if the user makes a request to /echo123, the server would first receive an upgrade OPTION request to /echo123, and then the actual user request (GET/POST/OPTIONS/DELETE, etc..) to /echo123.
After applying the change, the server will first receive an OPTION request to *, and then the actual user request to /echo123

Since we use R2 http2 only with Rest.li clients that implement the OPTIONS request, and we know the behavior of such a path, this didn't trigger any unwanted behaviors.
In a more generic use case (R2 client to non rest.li resources), this could have unintended consequences

Furthermore this change is needed to make the handshake before the first request comes in. In fact, before the first request comes in, we don't know the path.

Reference:
https://http2.github.io/faq/#can-i-implement-http2-without-implementing-http11

RB=1071804
BUG=SI-4006
G=si-core-reviewers
R=ssheng,bfeng,dhoa,cxu
A=ssheng