Enhance keyMapV3 to honor affinity routing configuration in d2.src
KeyMapper provides the APIs to map given (uri + keys) to a set of hosts in different partitions. The keys are primarily used to identify partitions. #mapKeyV3 also provide optional StickyKey argument to route the request to the same host in the same partition if their stickyKey are the same.

D2 config support affinity routing, which allows service to config uriRegex method, and extract stickyKey from URI through regex. This change enhances keyMapper#mapKeyV3 with this feature. Since #makeKeyV3 also supports stickyKey, it takes the precedence if the stickyKey is provided. If it is not provided, and the key can be extracted from URI according to d2config, we now honor this key as well.

ScatterGather class is also updated to use #mapKeyV3. #mapKeyV2 will be deprecated since V3 can cover all the cases used by V2.

RB=1171966
G=si-core-reviewers
R=xzhu,dhoa,fcapponi
A=xzhu