Releasing the pool the channel at shutdown to avoid deadlock until shutdownTimeout; Shutting down the SSL pool.
This (temporary) deadlock was probably hit very few times at shutdown until we had just one channelPoolManager. Since we had two channelPoolManagers for each client (one http and one https), this would become much more frequent. Luckily we had already in place a mechanism to avoid this indefinite deadlock, thanks to the shutdownTimeout (60s), which is working as an upper-bound and making this kind of deadlock negligible.

However it was not negligible for the test after we started shutting down the SSL pool, which is creating a race condition between shutting down and processing a channel. We should assure a best effort of the shutdown which should happen in few seconds (instead of just waiting for the timeout of 60seconds).

The deadlock would have happened just in the case of:
- an http channel was generated by the lifecycle
- concurrently the app was trying to shutdown and changing the client state

RB=1172849
G=si-core-reviewers
R=ssheng,dhoa,jgrande
A=ssheng