[pegasus] Emits R2 channel pool events
This change enables the ChannelPoolManagers to be registered against an EventProviderRegistry as a ChannelPoolEventProvider. ChannelPoolEventProvider supplies channel pool related metrics such as streaming, secure, http version, and PoolStats. The EventProviderRegistry is passed in externally and exposes a method to iterate every ChannelPoolEventProvider registered. The goal is to gather R2 level channel pool metrics event through the EventProviderRegistry.

I've also decided to bite the bullet to deprecate and refactor all the existing constructors in HttpClientFactory and create one private constructor only called by the builder. All existing use cases of the deprecated constructors are refactored to using the builder pattern.

RB=1278581
G=si-core-reviewers
R=dhoa,cxu,fcapponi
A=fcapponi