Support BannedUris property for cluster
'BannedUris' property is used to ban specific hosts for the given cluster. D2client will not pick the hosts given in the list. This property was moved to serviceProperties. We now add it back to cluster so user can ban the hosts at cluster level.

RB=1317765
G=si-core-reviewers
R=crzhang,ssheng,dhoa,fcapponi
A=crzhang,fcapponi