Fixed bug in ObjectIterator when using PRE_ORDER iteration order.
A typeref to an array would result in a DataElement with a null schema being passed to the callback for the array items.  All elements nested under that array typeref would also be passed a null schema.

This bug caused RestUtils.trimRecordTemplate() to incorrectly trim all the elements of the array.

RB=1386162
G=si-dev
R=gylin,kbalasub,aponniah,xma,mnchen
A=gylin,kbalasub,aponniah,xma,mnchen