Add "$returnEntity" query parameter to support returning the entity on demand
Added support for a new query parameter that allows clients to specify that they do or do not want
the entity to be returned for "return entity" methods. Also refactored IDL/snapshot generation tests,
and extended the tests to cover the "return entity" case.

RB=1393433
G=si-dev
R=kbalasub,xma,mlamure,mnchen
A=kbalasub,xma