Look up Javadoc by correct custom parameter name.
During IDL generation, the previous behavior is to look up Javadoc for
a method parameter with the same name as the query or action parameter
This changes the behavior to look up the name of the method parameter
corresponding to a query or action parameter and using that to look up
its Javadoc.

While rare and discouraged, we encounter legitimate situations in which
the name of a query or action parameter needs to be different from the
name of its corresponding Java method parameter. This happens primarily
as a consequence of migrations between code bases with different naming
conventions.

RB=1417423
G=si-core-reviewers
R=xma,mnchen
A=xma