Preserve order of sets added to request builders.
Prior to this change, sets added as query parameters to request builders
will be copied into a read-only HashSet before being sent over the wire.
This can change the iteration order unpredictably, even if the input set
is order-preserving (like LinkedHashSet or Guava's ImmutableSet), which
confuses tests and tooling expecting the input iteration order to be the
final output iteration order on the network.

RB=1418767
BUG=SI-7950,SI-7963
G=si-core-reviewers
R=wjohnson,xma,egokturk,mnchen
A=mnchen