Fix stale build directory for dataTemplates
This patch fixes DATAR-31. The tasks that stage PDSCs into the build directory for dataTemplate generation don't remove files from previous executions. This is problematic when renaming PDSCs, because the old PDSC files are left on the directory, sometimes creating problems for downstream tasks.
The solution is to remove the staging directory before executing any of the tasks that use it.

RB=1660568
R=sgizecki,mnchen
A=mnchen