Fix a race condition bug in CompositeWriter.
Race condition happens when the end of a entity stream in the composite
writer co-incides with wh.remaining() becoming zero.

Since the entity stream has reached the end, it will trigger onDone,
which in turn will trigger the initializtion of the next stream through
onInit.

Now, if the network stream invokes onWritePossibe (since previous
wh.remaining() returned zero), the execution in onWritePossible and
onInit will go in parallel, causing the incorrect handling of
_outstanding.

RB=1750381
G=sf-reviewers
R=seliang
A=seliang