Fix Gradle caching regression introduced in 27.3.8.
In Java plugins, Gradle cannot generate a consistent hash key for doFirst
and doLast actions attached to tasks, and so those tasks won't particpate
in up-to-date checking or caching.

In this commit, we wrap the doFirst and doLast actions for all tasks with
a CacheableAction class that Gradle will use to produce a consistent
cache key.

RB=1753446
G=si-dev
R=kbalasub,mnchen,evwillia,ymwang
A=ymwang