Streaming Timeout (Idle Timeout) Implementation - as detiled in the document @ https://docs.google.com/document/d/1s1dNjqoUkmo2TZ4mql4utHB2CZwe14JW8EPZFRmxOwA
Quick Summary of the problem statement: Currently in SI infrastructure we support Request Timeout - which is the total end to end duration of the request before the request timeout  is triggered. This works great if the end point methods are dealing with simple Restli resource methods where the request and response data size is consistent between requests. But when we do not know the exact size of the data being transmitted (like blobs) and  it varies between requests - it makes harder to configure an optimal request timeout value and the MP owners resort to a conservative configuration that satisfy the worst possible request timeout. This will introduce in-efficiency in the client side by failing to ‘fail fast’ and do not serve the purpose of protecting the client by timely short circuiting the connection.  It may also lead to bad user experience by not allowing the client side load balancer to find a healthy downstream alternative in a timely manner.

Implemented Solution: Introduce a new config ‘http.streamingTimeout’. The config ‘http.streamingTimeout’ is a client setting which sets the maximum inactivity time for an http streaming request. In other words, if a connection is open but no streaming activity is detected on  it for over idle-timeout time, the streamingTimeout exception will be raised. The inactivity can happen either in client side stream or in server side stream.

The existing http.requestTimeout - works the way it works currently. It Will timeout the request if the entire request/response do not complete within the configured ‘htttp.requestTimeout’ end to end.

RB=1794904
BUG=SI-12917
G=si-core-reviewers
R=ssheng,juncwang,fcapponi,cxu,evwillia
A=evwillia,juncwang,skaipa,xuli