Fix PDL encoder logic for writing import statements
Previously, the PDL encoder would write an import if it was outside the root
namespace of the schema being encoded. This breaks use cases with overridden
namespaces. This fix only writes imports for:
(1) Types outside the root namespace of the document
(2) Types declared outside the document

Also, this change adds more strict requirements in the PDL parser. The following
is prohibited:
(1) Importing types declared inside the document
(2) Importing types within the root namespace of the document
(3) Declaring types that conflict with existing imports

RB=1799751
G=sf-reviewers
R=mnchen,kbalasub
A=mnchen,kbalasub