Add varargs constructor to primitive arrays.
Following on the commit that generates varargs constructors for array
types, we also add the same to primitive arrays, namely:

* BooleanArray
* BytesArray
* DoubleArray
* FloatArray
* LongArray
* StringArray

We leave out IntegerArray on purpose because a constructor signature of

    public IntegerArray(Integer first, Integer... rest);

can be confused with:

    public IntegerArray(int initialCapacity);

when called with a single integer argument.

We also update code and tests to take advantage of the new constructors.

RB=1813571
BUG=SI-12989
G=si-dev
R=mnchen,kbalasub,evwillia,ssheng,bsoetarm
A=kbalasub