Support for data template schema field string literals in PDL
Encode the SCHEMA field in generated data templates according to the
file format of the source schema file. Postponing the use of a new
interface for files generated from PDSC to avoid build failures,
with the expectation that this will be addressed relatively soon.

RB=1864201
G=sf-reviewers
R=ybi,kbalasub,mnchen
A=ybi,kbalasub