Add support for framework latency instrumentation
As part of the SI instrumentation initiative, this commit adds timing
markers in a breadth of Rest.li-layer and R2-layer codepaths in order
to track framework latency. Client timings are temporarily disabled
for scatter-gather requests, and server timings are disabled for
multiplexed requests.

RB=1745918
G=sf-reviewers
R=ssheng,crzhang,bsoetarm,fcapponi,dmessink
A=crzhang,mnchen,kbalasub