Publish all schemas as-is, temporarily publish translated PDSC schemas
As of this commit, all schemas will now be published as-is in their respective
JARs, meaning that JARs may now include PDLs in addition to PDSCs. This commit
also publishes translated PDSC equivalents for all PDL schemas as a temporary
workaround for some use cases in the PDL migration, though this should be
removed once no longer needed. These translated schemas can be found in the
'legacyPegasusSchemas' directory in the root of the JAR.

RB=1943392
G=sf-reviewers
R=dhamilto,rjackson,kbalasub
A=kbalasub