Fix Maven publication for extra artifacts
The existing Maven publication setup wasn't including extra non-standard artifacts
such as fat JARs and testing JARs. This change resolves that and also manually adds
test dependency information into the POM. Ideally, the test artifacts shouldn't be
published, but for now this change is necessary to make it all work.