Protobuf parsing performance improvements. (#260)
Introduce support for ASCII only strings. Since this is a backward incompatible change, introduce a new content type for it, and deprecate the old one. By encoding ASCII only strings with a separate ordinal, we give a hint to decoders to use a faster path for decoding ASCII only strings in which bytes directly map to chars, and there are no multi-byte chars.

Co-authored-by: Karthik Ramgopal <kramgopa@linkedin.com>