Fix page load delays caused by search script
For some reason, the search script was enabled on all pages. This was an
issue because the search script loads up all textual content on the site
into memory. This was causing long page load times on the live website.
This addresses that by only enabling the search functionality on the
search page. It also cleans up the search scripts.