Implement Host Override List feature in D2.
Requirements

To support features like Fullstack Hovr, D2 needs to support the following routing override scenarios.
  1. Route requests going to a given cluster to a specific host.
  2. Route requests going to a given service to a specific host.
  3. Provide route overrides for multiple clusters and services.

D2 provides an existing override mechanism called TargetHints. However, TargetHints does not meet the above requirements due to the following restrictions.
TargetHints does not allow multiple overrides for different clusters and services.
TargetHints requires users to provide a full URL as opposed to a URL prefix like the D2 announcer. Therefore, we lose the flexibility of appending the service paths for different D2 services.

Implementation
Similar to TargetHints, HostOverrideList is also specified and passed through the RequestContext. HostOverrideList allows the user to add an arbitrary number of overrides for different clusters and services. When the request gets to the SimpleLoadBalancer, the implementation will look for the first matching override for the current cluster and service names. The request URI will be rewritten to be the override URI provided.
  1. hostOverrideList.addClusterOverride(“Cluster1”, “https://host:8080/context/”);
  2. hostOverrideList.addServiceOverride(“Service1”, “https://host:8080/context/”);

The provided override URI is a prefix, similar to the announced URI, as opposed to a full URI in the TargetHints mechanism. The service path will be appended to the end of the URI by the SimpleLoadBalancer when the request URI is rewritten. In the above examples, the written URI will be “https://host:8080/context/service1path”.

TargetHints mechanism will be deprecated. To achieve the original TargetHints behavior, one can add generic override without providing a cluster or service name.
hostOverrideList.addOverride(“https://host:8080/context/”);

Addressed Chao and Nizar's comments