Enable tests in Travis, add some exclusions
Enabling tests in the Travis CI build to give us higher confidence in
proposed changes. The Travis environment seems to introduce flakiness or
even total failure for some tests, so this adds test retries for some
tests and total exclusion for other tests in order to alleviate this.
The 'ci-flaky' test group is excluded for all test tasks in the CI
environment, though these tests should be addressed soon.

The process for addressing flakiness is:
- if a test is flaky, give it one retry using SingleRetry
- if it's still flaky, give it three retries using ThreeRetries
- if it's still flaky, disable it in CI using the group "ci-flaky"