Roll back Bintray upload on failure in Travis (#318)
There exists a problem now where a Bintray upload can partially fail,
leaving the repo in a bad state where future uploads may fail due to
conflicts. This fix adjusts the Travis script so that a failed Bintray
upload will result in an API call to completely delete the version.