Fix non-deterministic issue for annotations in the generated data template code by using right schema location for typeref class template (#323)
For data template code generation in rest.li, we annotate the generated data class with its original location file path. Internally, we use a stack to record the file paths when we traverse deeply into fields. This fix is to address the build inconsistency when the order of the data schema files are processed differently.