Update Pegasus Plugin's CopySchema tasks to delete stale schemas
The Issue:
The issue is that the pegasus plugin, when preparing the schemas for
publication, will copy the schemas to the build directory without
removing stale schemas. Without the fix, these stale schemas will be
packaged into the data template jar -> which is a problem.

The current workaround, which is implemented, is that we have a delete
task and a copy task. The delete task is triggered by a gradle project
flag. The copy moves the files as intended. This workaround is not
viable since it breaks the UP-TO-DATE checks between builds. In
particular, it breaks the UP-TO-DATE checks during a HOT RELOAD
(runPlay task of the Play Framework).

The Fix:
Instead of the workaround as above, we "skip" the delete task and make
the copy task become a sync task. The task name stays the same but the
class type will be a sync task which will "Synchronizes the contents of
a destination directory with some source directories and files
[https://docs.gradle.org/current/dsl/org.gradle.api.tasks.Sync.html]".
This task will now delete stale files in the destination and will pass
the 'up-to-date' between builds.

Testing:
Integ test:
1) Validate cacheability/up-to-date of tasks between builds
2) Validate the mainCopySchemas task will delete stale schemas

Unit test:
1) Validate the mainCopySchemas task is a Sync task

In addition, I worked with @DPUKyle to create multiple configurations
to allow the integ test to work. (Kyle's original PR:
https://github.com/Astro03/rest.li/pull/2)
 - Pass necessary configurations using sysprops and custom configurations
 - Custom task to keep pegasus-version.properties out of the integTest classpath
 - Override TemporaryFolder to prevent cleanup
 - Use settings.gradle to prevent randomly-generated values in project under test