Add new changelog (#333)
Replaces the old changelog with a new changelog conforming to the
"keep-a-changelog" standard, along with convenient links for seeing
which commits are included in each release. Adds a helper script for
updating the changelog and adding links as necessary. Also adds a check
in the Travis build script to ensure that it's updated.