Make addFields on AbstractRequestBuilder public.
We have a use case where we want a method to be able to take any type of
request builder (e.g. BatchGetRequestBuilder or FindRequestBuilder) and
we want to be able to add fields to the request.

Currently, we can accomplish this by using
setParam(RestConstants.FIELDS_PARAM, newHashSet(fields)), but it would
be simpler to just expose the addFields method instead.