Start adding suppport for multiple input dirs to pegasus (multiple data schema dirs).
This is useful for generating schemas. Our use case is for metadata (DataHub), we want to annotate metadata models with an annotation, and then auto generate event schemas based on those annotations (and then from there avro + java, etc). We want all models (hand written metadata models + auto generated events) in the same module.

Note that there seems to be a lack of tests in this area, I was able to add one but let me know if there are some other kind of tests or procedures I'm missing...