Fix the non-deterministic issue (#376)
The non-consistency problem happens in the following scenario:
Suppose there are 3 PDL files: 
* Foo1.pdl which has a field of array with type Bar 
* Foo2.pdl which also has a field of array with type Bar
This 3 schema file will generate Foo1.java Foo2.java and Bar.java and also BarArray.java

It is the annotation in the BarArray.java appeared to have non-consistency in terms appearing to have 2 different generated-by files. Sometimes it is Foo1.pdl, sometimes it is Foo2.pdl.

Solution is to have the input always sorted. This is way we take little hit on build-time performance, but we gain consistency where once we build, the following many times we can rely on build cache.