Fix bug referring to coercer before registration.
The new static DEFAULT_ fields that are code generated need to reference
coercers that might not be initialized until a subsequent static block.

All that is required to fix this is to reorder those statements, but it
does not seem possible with JCodeModel. What we can do instead is break
up the declaration from the initialization and order the assignment and
coercer registration statements the way we want to within a single init
block.

The generated code now looks like:

    private static final T DEFAULT_field;

    static {
      Custom.initializeCustomClass(T.class);
      Custom.initializeCoercerClass(...);
      DEFAULT_field = ...;
    }