Fix empty union validation to only allow validation pass when projection in union is present (#493)
Allow empty unions to pass when there is partial projections in the union. Internally, we will examine the projection during the request phase and when we are doing validation in the response phase, we will see if there exists a partial projection on the union schema. The flag `isPartialSchema `  is only used internally.