Sort properties and defaults when encoding PDLs.
The PDL encoder already sorts the top-level properties on a field by
their names. This extends this sorting to nested properties.

An example of a nested property that is sorted is:

    @nested = {
      "a" : "a",
      "b" : "b"
    }

When default values are records, their fields are sorted in the order
that the fields are defined in the schema. For example:

    date: record Date {
      month: int
      day: int
      year: int
    } = {
      "month": 1,
      "day": 1,
      "year": 2021
    }

This enables the PDL encoder to be used for autoformatting PDLs, since
the order of map entries will be deterministic.