Sort properties and defaults when encoding PDLs. (#515)
The PDL encoder already sorts the top-level properties on a field by
their names. This extends this sorting to nested JSON property values.

An example of a nested JSON property value that is sorted is:

    @nested = {
      "a" : "a",
      "b" : "b"
    }

For record field default values, the JSON objects are sorted in
the order that the fields are defined in the schema. For example:

    date: record Date {
      month: int
      day: int
      year: int
    } = {
      "month": 1,
      "day": 1,
      "year": 2021
    }

This enables the PDL encoder to be used for auto-formatting PDLs, since
the order of map entries will be deterministic.