Remove V1 only request and response classes.
This removes the following:

    BatchCreateRequest
    BatchCreateRequestBuilder
    CreateRequest
    CreateRequestBuilder

which are replaced by BatchCreateIdRequest and CreateIdRequest.

And removes:

    BatchGetRequest
    BatchGetKVRequest
    BatchRequestBuilder

which are replaced by BatchGetEntityRequest.

It also removes CreateResponse, which is now replaced by IdResponse,
as well as the deprecated Response#getId() method, since the ID can
now be retrieved from the IdResponse.

Finally, it also cleans up the code generators so that they no longer
can be configured to generate the V1 classes.